Move all internal/ packages to a root internal/
This will simplify my life a bit.
