+++ internal/and/and_closer.go
