+++ internal/compare/doc.go
