+++ internal/compare/image.go
