+++ internal/compare/image_test.go
