+++ internal/compare/index.go
