+++ internal/compare/index_test.go
