+++ internal/legacy/copy.go
