+++ pkg/v1/remote/transport/logger.go
@@ -20,7 +20,7 @@ import (
 	"net/http/httputil"
 	"time"
 
-	"github.com/google/go-containerregistry/pkg/internal/redact"
+	"github.com/google/go-containerregistry/internal/redact"
 	"github.com/google/go-containerregistry/pkg/logs"
 )
